requirements.txt
# PyPi publishing tools
setuptools
wheel
twine
pyinstaller
build
toml

# Secure File Storage
bcrypt
cryptography
loguru
pytest
flask
python-dotenv
